Split into scout_execute_ex, scout_internal_handlers and scout_observer separately
