+++ scout_utils.c
@@ -8,33 +8,6 @@
 #include "zend_scoutapm.h"
 #include "scout_extern.h"
 
-ZEND_NAMED_FUNCTION(scoutapm_default_handler);
-void add_function_to_instrumentation(const char *function_name);
-int should_be_instrumented(const char *function_name);
-const char *zval_type_and_value_if_possible(zval *val);
-
-void add_function_to_instrumentation(const char *function_name)
-{
-    if (SCOUTAPM_G(num_instrumented_functions) >= MAX_INSTRUMENTED_FUNCTIONS) {
-        zend_throw_exception_ex(NULL, 0, "Unable to add instrumentation for function '%s' - MAX_INSTRUMENTED_FUNCTIONS of %d reached", function_name, MAX_INSTRUMENTED_FUNCTIONS);
-        return;
-    }
-
-    SCOUTAPM_G(instrumented_function_names)[SCOUTAPM_G(num_instrumented_functions)] = strdup(function_name);
-    SCOUTAPM_G(num_instrumented_functions)++;
-}
-
-int should_be_instrumented(const char *function_name)
-{
-    int i = 0;
-    for (; i < SCOUTAPM_G(num_instrumented_functions); i++) {
-        if (strcasecmp(function_name, SCOUTAPM_G(instrumented_function_names)[i]) == 0) {
-            return 1;
-        }
-    }
-    return 0;
-}
-
 /*
  * Given some zend_execute_data, figure out what the function/method/static method is being called. The convention used
  * is `ClassName::methodName` for static methods, `ClassName->methodName` for instance methods, and `functionName` for
