+++ zend_scoutapm.h
@@ -16,6 +16,9 @@
 #include <zend_exceptions.h>
 #include "ext/standard/php_var.h"
 
+#include "scout_recording.h"
+#include "scout_internal_handlers.h"
+
 #define PHP_SCOUTAPM_NAME "scoutapm"
 #define PHP_SCOUTAPM_VERSION "1.2.2"
 
@@ -23,37 +26,15 @@
 #define SCOUT_APM_EXT_DEBUGGING 0
 
 #if PHP_VERSION_ID > 80000
-#define SCOUTAPM_INSTRUMENT_USING_OBSERVER_API 1
-#include "Zend/zend_observer.h"
+//#define SCOUTAPM_INSTRUMENT_USING_OBSERVER_API 1
+//#include "Zend/zend_observer.h"
+#define SCOUTAPM_INSTRUMENT_USING_OBSERVER_API 0
 #else
 #define SCOUTAPM_INSTRUMENT_USING_OBSERVER_API 0
 #endif
 
-
 PHP_FUNCTION(scoutapm_get_calls);
 PHP_FUNCTION(scoutapm_list_instrumented_functions);
-
-/* Describes information we store about a recorded stack frame */
-typedef struct _scoutapm_stack_frame {
-    const char *function_name;
-    double entered;
-    double exited;
-    int argc;
-    zval *argv;
-} scoutapm_stack_frame;
-
-typedef struct _scoutapm_disconnected_call_argument_store {
-    const char *reference;
-    int argc;
-    zval *argv;
-} scoutapm_disconnected_call_argument_store;
-
-typedef struct _handler_lookup {
-    int index;
-    const char *function_name;
-} indexed_handler_lookup;
-
-#define MAX_INSTRUMENTED_FUNCTIONS 100
 
 /* These are the "module globals". In non-ZTS mode, they're just regular variables, but means in ZTS mode they get handled properly */
 ZEND_BEGIN_MODULE_GLOBALS(scoutapm)
@@ -96,43 +77,6 @@ typedef void (*zif_handler)(INTERNAL_FUNCTION_PARAMETERS);
     sizeNeeded = snprintf(NULL, 0, fmt, ##__VA_ARGS__) + 1; \
     destString = (char*)malloc(sizeNeeded); \
     snprintf(destString, sizeNeeded, fmt, ##__VA_ARGS__)
-
-/* overload a regular function by wrapping its handler with our own handler */
-#define SCOUT_OVERLOAD_FUNCTION(function_name, handler_to_use) \
-    original_function = zend_hash_str_find_ptr(EG(function_table), function_name, sizeof(function_name) - 1); \
-    if (original_function != NULL) { \
-        handler_index = unchecked_handler_index_for_function(function_name); \
-        if (handler_index < 0) return FAILURE; \
-        original_handlers[handler_index] = original_function->internal_function.handler; \
-        original_function->internal_function.handler = handler_to_use; \
-    }
-
-/* Don't use this macro directly, use SCOUT_OVERLOAD_STATIC_METHOD or SCOUT_OVERLOAD_METHOD for consistency */
-#define SCOUT_OVERLOAD_CLASS_ENTRY_FUNCTION(lowercase_class_name, instance_or_static, method_name, handler_to_use) \
-    ce = zend_hash_str_find_ptr(CG(class_table), lowercase_class_name, sizeof(lowercase_class_name) - 1); \
-    if (ce != NULL) { \
-        original_function = zend_hash_str_find_ptr(&ce->function_table, method_name, sizeof(method_name)-1); \
-        if (original_function != NULL) { \
-            handler_index = unchecked_handler_index_for_function(lowercase_class_name instance_or_static method_name); \
-            if (handler_index < 0) return FAILURE; \
-            original_handlers[handler_index] = original_function->internal_function.handler; \
-            original_function->internal_function.handler = handler_to_use; \
-        } \
-    }
-
-/* overload a static class method by wrapping its handler with our own handler */
-#define SCOUT_OVERLOAD_STATIC_METHOD(lowercase_class_name, method_name, handler_to_use) SCOUT_OVERLOAD_CLASS_ENTRY_FUNCTION(lowercase_class_name, "::", method_name, handler_to_use)
-
-/* overload an instance class method by wrapping its handler with our own handler */
-#define SCOUT_OVERLOAD_METHOD(lowercase_class_name, method_name, handler_to_use) SCOUT_OVERLOAD_CLASS_ENTRY_FUNCTION(lowercase_class_name, "->", method_name, handler_to_use)
-
-#define SCOUT_INTERNAL_FUNCTION_PASSTHRU() original_handlers[handler_index_for_function(determine_function_name(execute_data))](INTERNAL_FUNCTION_PARAM_PASSTHRU)
-
-#define SCOUT_PASSTHRU_IF_ALREADY_INSTRUMENTING()   \
-    if (SCOUTAPM_G(currently_instrumenting) == 1) { \
-        SCOUT_INTERNAL_FUNCTION_PASSTHRU();         \
-        return;                                     \
-    }
 
 /* these are the string keys used in scoutapm_get_calls associative array return value */
 #define SCOUT_GET_CALLS_KEY_FUNCTION "function"
